+++ packages/hekireki/src/core/zod.ts
@@ -1,0 +1,39 @@
+import { mkdir, writeFile } from 'node:fs/promises'
+import path from 'node:path'
+
+import type { GeneratorOptions } from '@prisma/generator-helper'
+
+import { fmt } from '../format/index.js'
+import { makeRelationsOnly } from '../helper/prisma.js'
+import { makeZodRelations, zod as zodSchema } from '../helper/zod.js'
+import { getBool, getString } from '../utils/index.js'
+
+export async function zod(options: GeneratorOptions): Promise<void> {
+  if (!(options.generator.isCustomOutput && options.generator.output?.value)) {
+    throw new Error(
+      'output is required for Hekireki-Zod. Please specify output in your generator config.',
+    )
+  }
+  const output = options.generator.output.value
+  const resolved = path.extname(output)
+    ? { dir: path.dirname(output), file: output }
+    : { dir: output, file: path.join(output, 'index.ts') }
+  const zodVersion = getString(options.generator.config?.zod, 'v4')
+  const enableRelation = getBool(options.generator.config?.relation)
+
+  const base = zodSchema(
+    options.dmmf.datamodel.models,
+    getBool(options.generator.config?.type),
+    getBool(options.generator.config?.comment),
+    zodVersion,
+    options.dmmf.datamodel.enums,
+  )
+  const relations = enableRelation
+    ? makeRelationsOnly(options.dmmf, getBool(options.generator.config?.type), makeZodRelations)
+    : ''
+  const full = [base, relations].filter(Boolean).join('\n\n')
+
+  const code = await fmt(full)
+  await mkdir(resolved.dir, { recursive: true })
+  await writeFile(resolved.file, code, 'utf-8')
+}
