+++ packages/hekireki/src/generator/arktype/index.ts
@@ -1,48 +1,9 @@
 #!/usr/bin/env node
-import path from 'node:path'
-
-import type { GeneratorOptions } from '@prisma/generator-helper'
 import pkg from '@prisma/generator-helper'
 
-import { fmt } from '../../format/index.js'
-import { mkdir, writeFile } from '../../fsp/index.js'
-import { arktype, makeArktypeRelations } from '../../helper/arktype.js'
-import { makeRelationsOnly } from '../../helper/prisma.js'
-import { getBool } from '../../utils/index.js'
+import { arktype } from '../../core/arktype.js'
 
 const { generatorHandler } = pkg
-
-/**
- * Prisma generator entry point for ArkType schema generation
- * @param options - The Prisma generator options
- */
-export async function main(options: GeneratorOptions): Promise<void> {
-  if (!(options.generator.isCustomOutput && options.generator.output?.value)) {
-    throw new Error(
-      'output is required for Hekireki-ArkType. Please specify output in your generator config.',
-    )
-  }
-  const output = options.generator.output.value
-  const resolved = path.extname(output)
-    ? { dir: path.dirname(output), file: output }
-    : { dir: output, file: path.join(output, 'index.ts') }
-  const enableRelation = getBool(options.generator.config?.relation)
-
-  const base = arktype(
-    options.dmmf.datamodel.models,
-    getBool(options.generator.config?.type),
-    getBool(options.generator.config?.comment),
-    options.dmmf.datamodel.enums,
-  )
-  const relations = enableRelation
-    ? makeRelationsOnly(options.dmmf, getBool(options.generator.config?.type), makeArktypeRelations)
-    : ''
-  const full = [base, relations].filter(Boolean).join('\n\n')
-
-  const code = await fmt(full)
-  await mkdir(resolved.dir)
-  await writeFile(resolved.file, code)
-}
 
 generatorHandler({
   onManifest() {
@@ -51,5 +12,5 @@ generatorHandler({
       prettyName: 'Hekireki-ArkType',
     }
   },
-  onGenerate: main,
+  onGenerate: arktype,
 })
