+++ packages/hekireki/src/generator/ecto/index.ts
@@ -1,23 +1,9 @@
 #!/usr/bin/env node
-import type { GeneratorOptions } from '@prisma/generator-helper'
 import pkg from '@prisma/generator-helper'
 
-import { writeEctoSchemasToFiles } from '../../helper/ecto.js'
+import { ecto } from '../../core/ecto.js'
 
 const { generatorHandler } = pkg
-
-export async function main(options: GeneratorOptions): Promise<void> {
-  if (!(options.generator.isCustomOutput && options.generator.output?.value)) {
-    throw new Error(
-      'output is required for Hekireki-Ecto. Please specify output in your generator config.',
-    )
-  }
-  const output = options.generator.output.value
-  const app = options.generator.config?.app ?? 'MyApp'
-
-  const enums = options.dmmf.datamodel.enums
-  await writeEctoSchemasToFiles(options.dmmf.datamodel.models, app, output, enums)
-}
 
 generatorHandler({
   onManifest() {
@@ -26,5 +12,5 @@ generatorHandler({
       prettyName: 'Hekireki-Ecto',
     }
   },
-  onGenerate: main,
+  onGenerate: ecto,
 })
