+++ packages/hekireki/src/generator/gorm/index.ts
@@ -1,29 +1,9 @@
 #!/usr/bin/env node
-import path from 'node:path'
-
-import type { GeneratorOptions } from '@prisma/generator-helper'
 import pkg from '@prisma/generator-helper'
 
-import { writeGormFile } from '../../helper/gorm.js'
-import { getString } from '../../utils/index.js'
+import { gorm } from '../../core/gorm.js'
 
 const { generatorHandler } = pkg
-
-export async function main(options: GeneratorOptions): Promise<void> {
-  if (!(options.generator.isCustomOutput && options.generator.output?.value)) {
-    throw new Error(
-      'output is required for Hekireki-GORM. Please specify output in your generator config.',
-    )
-  }
-  const output = options.generator.output.value
-  const resolved = path.extname(output) ? output : path.join(output, 'models.go')
-
-  const packageName = getString(options.generator.config.package, 'model')
-
-  const enums = options.dmmf.datamodel.enums
-  const indexes = options.dmmf.datamodel.indexes
-  await writeGormFile(options.dmmf.datamodel.models, resolved, enums, indexes, packageName)
-}
 
 generatorHandler({
   onManifest() {
@@ -32,5 +12,5 @@ generatorHandler({
       prettyName: 'Hekireki-GORM',
     }
   },
-  onGenerate: main,
+  onGenerate: gorm,
 })
