+++ packages/hekireki/src/generator/sea-orm/index.ts
@@ -1,29 +1,9 @@
 #!/usr/bin/env node
-import type { GeneratorOptions } from '@prisma/generator-helper'
 import pkg from '@prisma/generator-helper'
 
-import type { SerdeOptions } from '../../helper/sea-orm.js'
-import { writeSeaOrmFiles } from '../../helper/sea-orm.js'
-import { getString } from '../../utils/index.js'
+import { seaOrm } from '../../core/sea-orm.js'
 
 const { generatorHandler } = pkg
-
-export async function main(options: GeneratorOptions): Promise<void> {
-  if (!(options.generator.isCustomOutput && options.generator.output?.value)) {
-    throw new Error(
-      'output is required for Hekireki-SeaORM. Please specify output in your generator config.',
-    )
-  }
-  const output = options.generator.output.value
-  const renameAll = getString(options.generator.config?.renameAll)
-
-  const serde: SerdeOptions = {
-    renameAll,
-  }
-
-  const enums = options.dmmf.datamodel.enums
-  await writeSeaOrmFiles(options.dmmf.datamodel.models, output, enums, serde)
-}
 
 generatorHandler({
   onManifest() {
@@ -32,5 +12,5 @@ generatorHandler({
       prettyName: 'Hekireki-SeaORM',
     }
   },
-  onGenerate: main,
+  onGenerate: seaOrm,
 })
