+++ packages/hekireki/src/generator/sqlalchemy/index.ts
@@ -1,26 +1,9 @@
 #!/usr/bin/env node
-import path from 'node:path'
-
-import type { GeneratorOptions } from '@prisma/generator-helper'
 import pkg from '@prisma/generator-helper'
 
-import { writeSQLAlchemyFile } from '../../helper/sqlalchemy.js'
+import { sqlalchemy } from '../../core/sqlalchemy.js'
 
 const { generatorHandler } = pkg
-
-export async function main(options: GeneratorOptions): Promise<void> {
-  if (!(options.generator.isCustomOutput && options.generator.output?.value)) {
-    throw new Error(
-      'output is required for Hekireki-SQLAlchemy. Please specify output in your generator config.',
-    )
-  }
-  const output = options.generator.output.value
-  const resolved = path.extname(output) ? output : path.join(output, 'models.py')
-
-  const enums = options.dmmf.datamodel.enums
-  const indexes = options.dmmf.datamodel.indexes
-  await writeSQLAlchemyFile(options.dmmf.datamodel.models, resolved, enums, indexes)
-}
 
 generatorHandler({
   onManifest() {
@@ -29,5 +12,5 @@ generatorHandler({
       prettyName: 'Hekireki-SQLAlchemy',
     }
   },
-  onGenerate: main,
+  onGenerate: sqlalchemy,
 })
